pyproject.toml
[tool.poetry]
authors = ["Camilo Velez <camilo.velez@axleinfo.com>"]
description = "ICT Python Client"
name = "ict"
packages = [{include = "ict", from = "src"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
cwl-utils = ">=0.30"
cwltool = "^3.1.20231020140205"
polus-plugins = {git = "https://github.com/PolusAI/image-tools"}
pydantic = {extras = ["email"], version = "^2.3"}
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
hypothesis = "^6.84.3"
nox = "^2023.4.22"
pytest = "^7.4.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
